+++ pkg/mpc/node.go
@@ -68,7 +68,6 @@ func NewNode(
 	}
 	logger.Info("Starting new node, preparams is generated successfully!")
 
-	peerRegistry.Ready()
 	go peerRegistry.WatchPeersReady()
 
 	return &Node{
@@ -141,7 +140,7 @@ func (p *Node) CreateSigningSession(
 	txID string,
 	networkInternalCode string,
 	threshold int,
-	successQueue messaging.MessageQueue,
+	resultQueue messaging.MessageQueue,
 ) (*SigningSession, error) {
 	readyPeerIDs := p.peerRegistry.GetReadyPeersIncludeSelf()
 	selfPartyID, allPartyIDs := p.generatePartyIDs(PurposeKeygen, readyPeerIDs)
@@ -158,7 +157,7 @@ func (p *Node) CreateSigningSession(
 		p.ecdsaPreParams,
 		p.kvstore,
 		p.keyinfoStore,
-		successQueue,
+		resultQueue,
 	)
 	return session, nil
 }
@@ -168,7 +167,7 @@ func (p *Node) CreateEDDSASigningSession(
 	txID string,
 	networkInternalCode string,
 	threshold int,
-	successQueue messaging.MessageQueue,
+	resultQueue messaging.MessageQueue,
 ) (*EDDSASigningSession, error) {
 	readyPeerIDs := p.peerRegistry.GetReadyPeersIncludeSelf()
 	selfPartyID, allPartyIDs := p.generatePartyIDs(PurposeKeygen, readyPeerIDs)
@@ -184,7 +183,7 @@ func (p *Node) CreateEDDSASigningSession(
 		threshold,
 		p.kvstore,
 		p.keyinfoStore,
-		successQueue,
+		resultQueue,
 	)
 	return session, nil
 }